LPS-159028 Move getIconSpriteMap() function
